requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dateparser==0.7.2
Flask==1.1.1
idna==2.8
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
requests==2.22.0
six==1.13.0
tqdm==4.38.0
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0
